#35890 Menu action fixes when multiple Max sessions are run at once.
The bug cropped up when multiple Max sessions are run concurrently. The macros attached to the menu actions changed after the second session's Shotgun menu wass built (Max caches the menu layout, including the macros, to disk) and the first session would call the macro defined in the second session. Because we keyed our cache of AppCommand objects off of the Python object ID, the lookup to find the AppCommand in the first session looked for the ID of the second session and fail.

Closes #10.