SG-12156 fix for getting parent window in 2018 - 2019 (#40)
This changes the way the engine attempts to get a parent widget in Max 2018 - 2019. The result should be that it's more stable/less error prone than the previous approach and we shouldn't see the `Could not start engine: 'NoneType' object has no attribute 'styleSheet'` anymore.

Also added additional logging around the bootstrap code to capture any errors with traceback and send them to the log.